add: exclude password and uuid parser pipe

+++ src/resources/user/user.controller.ts
@@ -6,6 +6,10 @@ import {
   Param,
   Delete,
   Put,
+  ParseUUIDPipe,
+  NotFoundException,
+  HttpStatus,
+  HttpCode,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -16,30 +20,58 @@ export class UserController {
   constructor(private readonly userService: UserService) {}
 
   @Post()
+  @HttpCode(HttpStatus.CREATED)
   create(@Body() createUserDto: CreateUserDto) {
-    return this.userService.create(createUserDto);
+    const { password, ...userWithoutPassword } =
+      this.userService.create(createUserDto);
+
+    return userWithoutPassword;
   }
 
   @Get()
   findAll() {
-    return this.userService.findAll();
+    const users = this.userService.findAll();
+
+    const withoutPasswordUsers = users.map(({ password, ...rest }) => rest);
+
+    return withoutPasswordUsers;
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.userService.findOne(id);
+  findOne(@Param('id', ParseUUIDPipe) id: string) {
+    const user = this.userService.findOne(id);
+
+    if (!user) throw new NotFoundException('User is not found!');
+
+    const { password, ...userWithoutPassword } = user;
+
+    return userWithoutPassword;
   }
 
   @Put(':id')
   update(
-    @Param('id') id: string,
+    @Param('id', ParseUUIDPipe) id: string,
     @Body() updatePasswordDto: UpdatePasswordDto,
   ) {
-    return this.userService.update(id, updatePasswordDto);
+    const user = this.userService.findOne(id);
+
+    if (!user) throw new NotFoundException('User is not found!');
+
+    const { password, ...userWithoutPassword } = this.userService.update(
+      id,
+      updatePasswordDto,
+    );
+
+    return userWithoutPassword;
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @HttpCode(HttpStatus.NO_CONTENT)
+  remove(@Param('id', ParseUUIDPipe) id: string) {
+    const user = this.userService.findOne(id);
+
+    if (!user) throw new NotFoundException('User is not found!');
+
     return this.userService.remove(id);
   }
 }
